Update dependencies [3/3]
Marked changed quite a bit, regarding the code changes:

`headerIds` and `headerPrefix` (replaced by the marked-gfm-heading-id package) no longer is necessary as the default is now for IDs _not_ to have IDs, which is what we want.

`mangle` (replaced by the marked-mangle package) no longer is necessary as well, as the default there too flipped from mangle to not mangle, which is the behaviour we want.