Add the ability to automatically generate string columns for date/time types
`ts-sql-codegen` generates our table structures, and we don't want to mess in the generated files. Similarly, we also don't want to subclass most of these generated classes, and using a utility function (`db.asDateTimeString()`) has proven a bit cumbersome as one manually has to keep track of column optionality.

This commit introduces a helper function, `withDateTimeStringColumns`, that automatically extends table definitions (& associated TypeScript type definitions) with string representations of DATE, DATETIME, TIME and TIMESTAMP columns. The representations will be created using the `DATE_FORMAT()` function, and are always forced to be in UTC.

One thing to note is that the generated string contains the "[UTC]" timezone suffix in addition to the regular ISO 8601 Z ("zulu") indicator. This is to maintain compatibility with the `Temporal.ZonedDateTime` factory functions, which require this.