Decrease TypeScript time by 51.3%
The `Theme` type exported from `@mui/system` is the "system theme", which the `Theme` type exported by `@mui/material` (that builds on top of the "system theme") is the one used by the components. In most cases this seems to go alright, however, there seem to be cases where TypeScript trips over the difference leading to a massive increase in type verification times.

The `tsc` compiler took 35.7 seconds to build this full repository before this change, and 17.4 seconds after this change, which makes this a 51.3% improvement.