Downgrade back to v14.1.4
As much fun as Next.js is to work with, stability most certainly isn't one of the strengths of the product. Today's entertainment brings "Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: object" errors on loads of pages in our app in production, none of which occur in the dev server.