feat: Add a modifier key to only drag shifts between volunteers
It can be tedious to move a shift to a volunteer, and to accidentally move it to a different time as well. This PR pulls in v1.8.0 of our timeline component, which will restrict dragging shifts to different resources, but not to different times, when the "shift" key is pressed.